feat: :recycle: move config .env & create database file
