+++ src/database/entities/ad.entity.ts
