+++ src/database/entities/viewRecord.entity.ts
