+++ src/modules/ads/ads.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { AdsController } from './ads.controller';
 import { AdsService } from './ads.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Ad } from 'src/entities/ad.entity';
+import { Ad } from 'src/database/entities/ad.entity';
 
-import { ViewRecord } from 'src/entities/viewRecord.entity';
-import { WebpageOwner } from 'src/entities/webpageOwner.entity';
-import { WebpageOwnerView } from 'src/entities/pageOwnerView.entity';
+import { ViewRecord } from 'src/database/entities/viewRecord.entity';
+import { WebpageOwner } from 'src/database/entities/webpageOwner.entity';
+import { WebpageOwnerView } from 'src/database/entities/pageOwnerView.entity';
 
 @Module({
   imports: [
