+++ src/modules/ads/ads.service.ts
@@ -1,15 +1,15 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { ViewAdsDTO } from './dto/createAds.dto';
 import { Repository } from 'typeorm';
-import { Ad } from 'src/entities/ad.entity';
+import { Ad } from 'src/database/entities/ad.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JsonRpcProvider, ethers } from 'ethers';
 import { Cron } from '@nestjs/schedule';
 
-import { ViewRecord } from 'src/entities/viewRecord.entity';
+import { ViewRecord } from 'src/database/entities/viewRecord.entity';
 import * as dayjs from 'dayjs';
-import { WebpageOwner } from 'src/entities/webpageOwner.entity';
-import { WebpageOwnerView } from 'src/entities/pageOwnerView.entity';
+import { WebpageOwner } from 'src/database/entities/webpageOwner.entity';
+import { WebpageOwnerView } from 'src/database/entities/pageOwnerView.entity';
 import { BillBOSCore_ABI } from 'constants/abis/billbos-core.abi';
 
 @Injectable()
